Remove old ContactDetails component
